requirements.txt
PySide6-Essentials>=6.6
pyinstaller>=6.0
